requirements.txt
tensorflow
keras
Django
djangorestframework
drf-yasg
numpy
pandas
sklearn
gunicorn
pillow
polls
markdown
django-filter
pyrebase4